pyproject.toml
[project]
name = "cs224n"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "datasets>=2.18.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "modal",
]
